# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Student API",
    "version": "1.0.0",
    "description": "API documentation for managing students and their vaccination details."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local server"
    }
  ],
  "paths": {
    "/students": {
      "get": {
        "summary": "Get all students",
        "description": "Fetches all students from the database.",
        "responses": {
          "200": {
            "description": "A list of students",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Student"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new student",
        "description": "Adds a new student to the database.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Student created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          }
        }
      }
    },
    "/students/{id}": {
      "get": {
        "summary": "Get a student by ID",
        "description": "Fetches a student by their ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the student to fetch."
          }
        ],
        "responses": {
          "200": {
            "description": "Student details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "put": {
        "summary": "Update a student",
        "description": "Updates a student's details.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the student to update."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Student"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a student",
        "description": "Deletes a student by their ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the student to delete."
          }
        ],
        "responses": {
          "200": {
            "description": "Student deleted successfully"
          },
          "404": {
            "description": "Student not found"
          }
        }
      }
    },
    "/students/{id}/vaccinate": {
      "post": {
        "summary": "Vaccinate a student",
        "description": "Marks a student as vaccinated.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the student to vaccinate."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "vaccineName": {
                    "type": "string",
                    "description": "The name of the vaccine."
                  },
                  "date": {
                    "type": "string",
                    "format": "date",
                    "description": "The date of vaccination."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Student vaccinated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Student"
                }
              }
            }
          },
          "404": {
            "description": "Student not found"
          }
        }
      }
    },
    "/reports": {
      "get": {
        "summary": "Get vaccination reports",
        "description": "Fetches vaccination reports based on query parameters.",
        "parameters": [
          {
            "name": "vaccineName",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Filter reports by vaccine name."
          }
        ],
        "responses": {
          "200": {
            "description": "A list of vaccination reports",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "studentName": { "type": "string" },
                      "className": { "type": "string" },
                      "status": { "type": "string" },
                      "date": { "type": "string", "format": "date" },
                      "vaccineName": { "type": "string" }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Student": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "age": {
            "type": "integer"
          },
          "class": {
            "type": "string"
          },
          "vaccinated": {
            "type": "boolean"
          },
          "vaccinationDetails": {
            "type": "object",
            "properties": {
              "vaccineName": {
                "type": "string"
              },
              "date": {
                "type": "string",
                "format": "date"
              }
            }
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}